Apply C++ standard compliance for cross-platform G++ build (#1990)
* Apply C++ standard compliance for cross-platform G++ build

* Fix include path case to match filesystem for Linux build

* Replace platform-dependent long with portable types

* Address review: Ammo::Clamp on int, FlareControl early return

* Update ShaderManager.cpp

---------

+++ TombEngine/Game/control/box.cpp
@@ -3346,7 +3346,7 @@ void InitializeItemBoxData()
 	{
 		for (const auto& mesh : room.mesh)
 		{
-			long index = ((mesh.Pose.Position.z - room.Position.z) / BLOCK(1)) + room.ZSize * ((mesh.Pose.Position.x - room.Position.x) / BLOCK(1));
+			int index = ((mesh.Pose.Position.z - room.Position.z) / BLOCK(1)) + room.ZSize * ((mesh.Pose.Position.x - room.Position.x) / BLOCK(1));
 			if (index >= room.Sectors.size())
 				continue;
